Created new file mcq.css to style the mcq container in mcq.html page. Created variable 'runningQuestionIndex' to store the current question index. With help of these variable we render the mcq question on page. To check the mcq answer is right or wrong we have created function 'checKAnswer()'.
Function:
renderQuestion() - It render the mcq , when answer is provided
scoreSheet() - It visible div when mcq is completed or time is completed.
redirectTOHighscorePage() - To redirect to highscore page when click on
submit. (Not working properly, it is redirecting to first mcq page)
checkAnswer() - When click on any option the function will be called and
give output as incorrect or correct.
showdiv() - It visible the horizontal line and answer container.
unshowdiv() - It hide the horizontal line and answer container.

Link to style.css, mcq.css and mcq.js.
./assets/css/style.css -> for styling nav bar
./assets/css/mcq.css -> for styling mcq.html page
assets/mcq.js -> for making mcq.html page responsive

In header section we have created nav bar with two childs 'Highscore'
and 'Time'. And in main section mcq container is created with
dynamically rendering.